gleam::ffi_gl::__gl_imports::mem

Function transmute_neo

pub const unsafe fn transmute_neo<Src, Dst>(src: Src) -> Dst

🔬This is a nightly-only experimental API. (transmute_neo)

New version of transmute, exposed under this name so it can be iterated upon without risking breakage to uses of “real” transmute.

It will not be stabilized under this name.

Examples

#![feature(transmute_neo)]
use std::mem::transmute_neo;

assert_eq!(unsafe { transmute_neo::<f32, u32>(0.0) }, 0);

#![feature(transmute_neo)]
use std::mem::transmute_neo;

unsafe { transmute_neo::<u32, u16>(123) };