gleam::ffi_gles::__gl_imports::raw

Type Alias c_double
·

pub type c_double = f64;

Equivalent to C’s double type.

This type will almost always be f64, which is guaranteed to be an IEEE 754 double-precision float in Rust. That said, the standard technically only guarantees that it be a floating-point number with at least the precision of a float; some 16-bit systems use f32, for example. Esoteric systems could use something entirely different from the IEEE-754 standard.