Struct gleam::gl::ProfilingGl

pub struct ProfilingGl<F> {
    gl: Rc<dyn Gl>,
    threshold: Duration,
    callback: F,
}

A wrapper around GL context that times each call and invokes the callback if the call takes longer than the threshold.

Fields
gl: Rc<dyn Gl>threshold: Durationcallback: F
Implementations

impl<F: 'static + Fn(&str, Duration)> ProfilingGl<F>

pub fn wrap(fns: Rc<dyn Gl>, threshold: Duration, callback: F) -> Rc<dyn Gl>