Module gleam::gl

Modules

types

Macros

declare_gl_apis 🔒

Structs

DebugMessage

ErrorCheckingGl

ErrorReactingGl

A wrapper around GL context that calls a specified callback on each GL error.

FnPtr

Gl

GlFfi

GlFns

GlesFfi

GlesFns

ProfilingGl

A wrapper around GL context that times each call and invokes the callback if the call takes longer than the threshold.

Enums

GlType

__GLsync

_cl_context

_cl_event

Constants

ACCUM

ACCUM_ALPHA_BITS

ACCUM_BLUE_BITS

ACCUM_BUFFER_BIT

ACCUM_CLEAR_VALUE

ACCUM_GREEN_BITS

ACCUM_RED_BITS

ACTIVE_ATOMIC_COUNTER_BUFFERS

ACTIVE_ATTRIBUTES

ACTIVE_ATTRIBUTE_MAX_LENGTH

ACTIVE_PROGRAM

ACTIVE_RESOURCES

ACTIVE_TEXTURE

ACTIVE_UNIFORMS

ACTIVE_UNIFORM_BLOCKS

ACTIVE_UNIFORM_BLOCK_MAX_NAME_LENGTH

ACTIVE_UNIFORM_MAX_LENGTH

ACTIVE_VARIABLES

ADD

ADD_SIGNED

ALIASED_LINE_WIDTH_RANGE

ALIASED_POINT_SIZE_RANGE

ALL_ATTRIB_BITS

ALL_BARRIER_BITS

ALL_SHADER_BITS

ALPHA

ALPHA4

ALPHA8

ALPHA8_EXT

ALPHA12

ALPHA16

ALPHA16F_EXT

ALPHA32F_EXT

ALPHA_BIAS

ALPHA_BITS

ALPHA_INTEGER

ALPHA_SCALE

ALPHA_TEST

ALPHA_TEST_FUNC

ALPHA_TEST_REF

ALREADY_SIGNALED

ALWAYS

AMBIENT

AMBIENT_AND_DIFFUSE

AND

AND_INVERTED

AND_REVERSE

ANY_SAMPLES_PASSED

ANY_SAMPLES_PASSED_CONSERVATIVE

ARRAY_BUFFER

ARRAY_BUFFER_BINDING

ARRAY_SIZE

ARRAY_STRIDE

ATOMIC_COUNTER_BARRIER_BIT

ATOMIC_COUNTER_BUFFER

ATOMIC_COUNTER_BUFFER_BINDING

ATOMIC_COUNTER_BUFFER_INDEX

ATOMIC_COUNTER_BUFFER_SIZE

ATOMIC_COUNTER_BUFFER_START

ATTACHED_SHADERS

ATTRIB_STACK_DEPTH

AUTO_NORMAL

AUX0

AUX1

AUX2

AUX3

AUX_BUFFERS

BACK

BACK_LEFT

BACK_RIGHT

BGR

BGRA

BGRA8_EXT

BGRA_EXT

BGRA_INTEGER

BGR_INTEGER

BITMAP

BITMAP_TOKEN

BLEND

BLEND_ADVANCED_COHERENT_KHR

BLEND_COLOR

BLEND_DST

BLEND_DST_ALPHA

BLEND_DST_RGB

BLEND_EQUATION

BLEND_EQUATION_ALPHA

BLEND_EQUATION_RGB

BLEND_SRC

BLEND_SRC_ALPHA

BLEND_SRC_RGB

BLOCK_INDEX

BLUE

BLUE_BIAS

BLUE_BITS

BLUE_INTEGER

BLUE_SCALE

BOOL

BOOL_VEC2

BOOL_VEC3

BOOL_VEC4

BUFFER

BUFFER_ACCESS

BUFFER_ACCESS_FLAGS

BUFFER_BINDING

BUFFER_DATA_SIZE

BUFFER_IMMUTABLE_STORAGE

BUFFER_IMMUTABLE_STORAGE_EXT

BUFFER_KHR

BUFFER_MAPPED

BUFFER_MAP_LENGTH

BUFFER_MAP_OFFSET

BUFFER_MAP_POINTER

BUFFER_SIZE

BUFFER_STORAGE_FLAGS

BUFFER_STORAGE_FLAGS_EXT

BUFFER_UPDATE_BARRIER_BIT

BUFFER_USAGE

BUFFER_VARIABLE

BYTE

C3F_V3F

C4F_N3F_V3F

C4UB_V2F

C4UB_V3F

CCW

CLAMP

CLAMP_FRAGMENT_COLOR

CLAMP_READ_COLOR

CLAMP_TO_BORDER

CLAMP_TO_EDGE

CLAMP_VERTEX_COLOR

CLEAR

CLIENT_ACTIVE_TEXTURE

CLIENT_ALL_ATTRIB_BITS

CLIENT_ATTRIB_STACK_DEPTH

CLIENT_MAPPED_BUFFER_BARRIER_BIT

CLIENT_MAPPED_BUFFER_BARRIER_BIT_EXT

CLIENT_PIXEL_STORE_BIT

CLIENT_STORAGE_BIT

CLIENT_STORAGE_BIT_EXT

CLIENT_VERTEX_ARRAY_BIT

CLIP_DISTANCE0

CLIP_DISTANCE1

CLIP_DISTANCE2

CLIP_DISTANCE3

CLIP_DISTANCE4

CLIP_DISTANCE5

CLIP_DISTANCE6

CLIP_DISTANCE7

CLIP_PLANE0

CLIP_PLANE1

CLIP_PLANE2

CLIP_PLANE3

CLIP_PLANE4

CLIP_PLANE5

COEFF

COLOR

COLORBURN_KHR

COLORDODGE_KHR

COLOR_ARRAY

COLOR_ARRAY_BUFFER_BINDING

COLOR_ARRAY_POINTER

COLOR_ARRAY_SIZE

COLOR_ARRAY_STRIDE

COLOR_ARRAY_TYPE

COLOR_ATTACHMENT0

COLOR_ATTACHMENT1

COLOR_ATTACHMENT2

COLOR_ATTACHMENT3

COLOR_ATTACHMENT4

COLOR_ATTACHMENT5

COLOR_ATTACHMENT6

COLOR_ATTACHMENT7

COLOR_ATTACHMENT8

COLOR_ATTACHMENT9

COLOR_ATTACHMENT10

COLOR_ATTACHMENT11

COLOR_ATTACHMENT12

COLOR_ATTACHMENT13

COLOR_ATTACHMENT14

COLOR_ATTACHMENT15

COLOR_ATTACHMENT16

COLOR_ATTACHMENT17

COLOR_ATTACHMENT18

COLOR_ATTACHMENT19

COLOR_ATTACHMENT20

COLOR_ATTACHMENT21

COLOR_ATTACHMENT22

COLOR_ATTACHMENT23

COLOR_ATTACHMENT24

COLOR_ATTACHMENT25

COLOR_ATTACHMENT26

COLOR_ATTACHMENT27

COLOR_ATTACHMENT28

COLOR_ATTACHMENT29

COLOR_ATTACHMENT30

COLOR_ATTACHMENT31

COLOR_BUFFER_BIT

COLOR_BUFFER_BIT0_QCOM

COLOR_BUFFER_BIT1_QCOM

COLOR_BUFFER_BIT2_QCOM

COLOR_BUFFER_BIT3_QCOM

COLOR_BUFFER_BIT4_QCOM

COLOR_BUFFER_BIT5_QCOM

COLOR_BUFFER_BIT6_QCOM

COLOR_BUFFER_BIT7_QCOM

COLOR_CLEAR_VALUE

COLOR_INDEX

COLOR_INDEXES

COLOR_LOGIC_OP

COLOR_MATERIAL

COLOR_MATERIAL_FACE

COLOR_MATERIAL_PARAMETER

COLOR_SUM

COLOR_WRITEMASK

COMBINE

COMBINE_ALPHA

COMBINE_RGB

COMMAND_BARRIER_BIT

COMPARE_REF_TO_TEXTURE

COMPARE_R_TO_TEXTURE

COMPILE

COMPILE_AND_EXECUTE

COMPILE_STATUS

COMPRESSED_ALPHA

COMPRESSED_INTENSITY

COMPRESSED_LUMINANCE

COMPRESSED_LUMINANCE_ALPHA

COMPRESSED_R11_EAC

COMPRESSED_RED

COMPRESSED_RED_RGTC1

COMPRESSED_RG

COMPRESSED_RG11_EAC

COMPRESSED_RGB

COMPRESSED_RGB8_ETC2

COMPRESSED_RGB8_PUNCHTHROUGH_ALPHA1_ETC2

COMPRESSED_RGBA

COMPRESSED_RGBA8_ETC2_EAC

COMPRESSED_RG_RGTC2

COMPRESSED_SIGNED_R11_EAC

COMPRESSED_SIGNED_RED_RGTC1

COMPRESSED_SIGNED_RG11_EAC

COMPRESSED_SIGNED_RG_RGTC2

COMPRESSED_SLUMINANCE

COMPRESSED_SLUMINANCE_ALPHA

COMPRESSED_SRGB

COMPRESSED_SRGB8_ALPHA8_ETC2_EAC

COMPRESSED_SRGB8_ETC2

COMPRESSED_SRGB8_PUNCHTHROUGH_ALPHA1_ETC2

COMPRESSED_SRGB_ALPHA

COMPRESSED_TEXTURE_FORMATS

COMPUTE_SHADER

COMPUTE_SHADER_BIT

COMPUTE_WORK_GROUP_SIZE

CONDITION_SATISFIED

CONSTANT

CONSTANT_ALPHA

CONSTANT_ATTENUATION

CONSTANT_COLOR

CONTEXT_COMPATIBILITY_PROFILE_BIT

CONTEXT_CORE_PROFILE_BIT

CONTEXT_FLAGS

CONTEXT_FLAG_DEBUG_BIT

CONTEXT_FLAG_DEBUG_BIT_KHR

CONTEXT_FLAG_FORWARD_COMPATIBLE_BIT

CONTEXT_PROFILE_MASK

COORD_REPLACE

COPY

COPY_INVERTED

COPY_PIXEL_TOKEN

COPY_READ_BUFFER

COPY_READ_BUFFER_BINDING

COPY_WRITE_BUFFER

COPY_WRITE_BUFFER_BINDING

CULL_FACE

CULL_FACE_MODE

CURRENT_BIT

CURRENT_COLOR

CURRENT_FOG_COORD

CURRENT_FOG_COORDINATE

CURRENT_INDEX

CURRENT_NORMAL

CURRENT_PROGRAM

CURRENT_QUERY

CURRENT_QUERY_EXT

CURRENT_RASTER_COLOR

CURRENT_RASTER_DISTANCE

CURRENT_RASTER_INDEX

CURRENT_RASTER_POSITION

CURRENT_RASTER_POSITION_VALID

CURRENT_RASTER_SECONDARY_COLOR

CURRENT_RASTER_TEXTURE_COORDS

CURRENT_SECONDARY_COLOR

CURRENT_TEXTURE_COORDS

CURRENT_VERTEX_ATTRIB

CW

DARKEN_KHR

DEBUG_CALLBACK_FUNCTION

DEBUG_CALLBACK_FUNCTION_KHR

DEBUG_CALLBACK_USER_PARAM

DEBUG_CALLBACK_USER_PARAM_KHR

DEBUG_GROUP_STACK_DEPTH

DEBUG_GROUP_STACK_DEPTH_KHR

DEBUG_LOGGED_MESSAGES

DEBUG_LOGGED_MESSAGES_KHR

DEBUG_NEXT_LOGGED_MESSAGE_LENGTH

DEBUG_NEXT_LOGGED_MESSAGE_LENGTH_KHR

DEBUG_OUTPUT

DEBUG_OUTPUT_KHR

DEBUG_OUTPUT_SYNCHRONOUS

DEBUG_OUTPUT_SYNCHRONOUS_KHR

DEBUG_SEVERITY_HIGH

DEBUG_SEVERITY_HIGH_KHR

DEBUG_SEVERITY_LOW

DEBUG_SEVERITY_LOW_KHR

DEBUG_SEVERITY_MEDIUM

DEBUG_SEVERITY_MEDIUM_KHR

DEBUG_SEVERITY_NOTIFICATION

DEBUG_SEVERITY_NOTIFICATION_KHR

DEBUG_SOURCE_API

DEBUG_SOURCE_API_KHR

DEBUG_SOURCE_APPLICATION

DEBUG_SOURCE_APPLICATION_KHR

DEBUG_SOURCE_OTHER

DEBUG_SOURCE_OTHER_KHR

DEBUG_SOURCE_SHADER_COMPILER

DEBUG_SOURCE_SHADER_COMPILER_KHR

DEBUG_SOURCE_THIRD_PARTY

DEBUG_SOURCE_THIRD_PARTY_KHR

DEBUG_SOURCE_WINDOW_SYSTEM

DEBUG_SOURCE_WINDOW_SYSTEM_KHR

DEBUG_TYPE_DEPRECATED_BEHAVIOR

DEBUG_TYPE_DEPRECATED_BEHAVIOR_KHR

DEBUG_TYPE_ERROR

DEBUG_TYPE_ERROR_KHR

DEBUG_TYPE_MARKER

DEBUG_TYPE_MARKER_KHR

DEBUG_TYPE_OTHER

DEBUG_TYPE_OTHER_KHR

DEBUG_TYPE_PERFORMANCE

DEBUG_TYPE_PERFORMANCE_KHR

DEBUG_TYPE_POP_GROUP

DEBUG_TYPE_POP_GROUP_KHR

DEBUG_TYPE_PORTABILITY

DEBUG_TYPE_PORTABILITY_KHR

DEBUG_TYPE_PUSH_GROUP

DEBUG_TYPE_PUSH_GROUP_KHR

DEBUG_TYPE_UNDEFINED_BEHAVIOR

DEBUG_TYPE_UNDEFINED_BEHAVIOR_KHR

DECAL

DECR

DECR_WRAP

DELETE_STATUS

DEPTH

DEPTH24_STENCIL8

DEPTH32F_STENCIL8

DEPTH_ATTACHMENT

DEPTH_BIAS

DEPTH_BITS

DEPTH_BUFFER_BIT

DEPTH_BUFFER_BIT0_QCOM

DEPTH_BUFFER_BIT1_QCOM

DEPTH_BUFFER_BIT2_QCOM

DEPTH_BUFFER_BIT3_QCOM

DEPTH_BUFFER_BIT4_QCOM

DEPTH_BUFFER_BIT5_QCOM

DEPTH_BUFFER_BIT6_QCOM

DEPTH_BUFFER_BIT7_QCOM

DEPTH_CLAMP

DEPTH_CLEAR_VALUE

DEPTH_COMPONENT

DEPTH_COMPONENT16

DEPTH_COMPONENT24

DEPTH_COMPONENT32

DEPTH_COMPONENT32F

DEPTH_FUNC

DEPTH_RANGE

DEPTH_SCALE

DEPTH_STENCIL

DEPTH_STENCIL_ATTACHMENT

DEPTH_STENCIL_TEXTURE_MODE

DEPTH_TEST

DEPTH_TEXTURE_MODE

DEPTH_WRITEMASK

DIFFERENCE_KHR

DIFFUSE

DISPATCH_INDIRECT_BUFFER

DISPATCH_INDIRECT_BUFFER_BINDING

DISPLAY_LIST

DITHER

DOMAIN

DONT_CARE

DOT3_RGB

DOT3_RGBA

DOUBLE

DOUBLEBUFFER

DRAW_BUFFER

DRAW_BUFFER0

DRAW_BUFFER1

DRAW_BUFFER2

DRAW_BUFFER3

DRAW_BUFFER4

DRAW_BUFFER5

DRAW_BUFFER6

DRAW_BUFFER7

DRAW_BUFFER8

DRAW_BUFFER9

DRAW_BUFFER10

DRAW_BUFFER11

DRAW_BUFFER12

DRAW_BUFFER13

DRAW_BUFFER14

DRAW_BUFFER15

DRAW_FRAMEBUFFER

DRAW_FRAMEBUFFER_BINDING

DRAW_INDIRECT_BUFFER

DRAW_INDIRECT_BUFFER_BINDING

DRAW_PIXELS_APPLE

DRAW_PIXEL_TOKEN

DST_ALPHA

DST_COLOR

DYNAMIC_COPY

DYNAMIC_DRAW

DYNAMIC_READ

DYNAMIC_STORAGE_BIT

DYNAMIC_STORAGE_BIT_EXT

EDGE_FLAG

EDGE_FLAG_ARRAY

EDGE_FLAG_ARRAY_BUFFER_BINDING

EDGE_FLAG_ARRAY_POINTER

EDGE_FLAG_ARRAY_STRIDE

ELEMENT_ARRAY_BARRIER_BIT

ELEMENT_ARRAY_BUFFER

ELEMENT_ARRAY_BUFFER_BINDING

EMISSION

ENABLE_BIT

EQUAL

EQUIV

EVAL_BIT

EXCLUSION_KHR

EXP

EXP2

EXTENSIONS

EYE_LINEAR

EYE_PLANE

FALSE

FASTEST

FEEDBACK

FEEDBACK_BUFFER_POINTER

FEEDBACK_BUFFER_SIZE

FEEDBACK_BUFFER_TYPE

FENCE_APPLE

FILL

FIRST_VERTEX_CONVENTION

FIXED

FIXED_ONLY

FLAT

FLOAT

FLOAT_32_UNSIGNED_INT_24_8_REV

FLOAT_MAT2

FLOAT_MAT2x3

FLOAT_MAT2x4

FLOAT_MAT3

FLOAT_MAT3x2

FLOAT_MAT3x4

FLOAT_MAT4

FLOAT_MAT4x2

FLOAT_MAT4x3

FLOAT_VEC2

FLOAT_VEC3

FLOAT_VEC4

FOG

FOG_BIT

FOG_COLOR

FOG_COORD

FOG_COORDINATE

FOG_COORDINATE_ARRAY

FOG_COORDINATE_ARRAY_BUFFER_BINDING

FOG_COORDINATE_ARRAY_POINTER

FOG_COORDINATE_ARRAY_STRIDE

FOG_COORDINATE_ARRAY_TYPE

FOG_COORDINATE_SOURCE

FOG_COORD_ARRAY

FOG_COORD_ARRAY_BUFFER_BINDING

FOG_COORD_ARRAY_POINTER

FOG_COORD_ARRAY_STRIDE

FOG_COORD_ARRAY_TYPE

FOG_COORD_SRC

FOG_DENSITY

FOG_END

FOG_HINT

FOG_INDEX

FOG_MODE

FOG_START

FRAGMENT_DEPTH

FRAGMENT_SHADER

FRAGMENT_SHADER_BIT

FRAGMENT_SHADER_DERIVATIVE_HINT

FRAMEBUFFER

FRAMEBUFFER_ATTACHMENT_ALPHA_SIZE

FRAMEBUFFER_ATTACHMENT_ANGLE

FRAMEBUFFER_ATTACHMENT_BLUE_SIZE

FRAMEBUFFER_ATTACHMENT_COLOR_ENCODING

FRAMEBUFFER_ATTACHMENT_COMPONENT_TYPE

FRAMEBUFFER_ATTACHMENT_DEPTH_SIZE

FRAMEBUFFER_ATTACHMENT_GREEN_SIZE

FRAMEBUFFER_ATTACHMENT_LAYERED

FRAMEBUFFER_ATTACHMENT_OBJECT_NAME

FRAMEBUFFER_ATTACHMENT_OBJECT_TYPE

FRAMEBUFFER_ATTACHMENT_RED_SIZE

FRAMEBUFFER_ATTACHMENT_STENCIL_SIZE

FRAMEBUFFER_ATTACHMENT_TEXTURE_CUBE_MAP_FACE

FRAMEBUFFER_ATTACHMENT_TEXTURE_LAYER

FRAMEBUFFER_ATTACHMENT_TEXTURE_LEVEL

FRAMEBUFFER_BARRIER_BIT

FRAMEBUFFER_BINDING

FRAMEBUFFER_COMPLETE

FRAMEBUFFER_DEFAULT

FRAMEBUFFER_DEFAULT_FIXED_SAMPLE_LOCATIONS

FRAMEBUFFER_DEFAULT_HEIGHT

FRAMEBUFFER_DEFAULT_SAMPLES

FRAMEBUFFER_DEFAULT_WIDTH

FRAMEBUFFER_INCOMPLETE_ATTACHMENT

FRAMEBUFFER_INCOMPLETE_DIMENSIONS

FRAMEBUFFER_INCOMPLETE_DRAW_BUFFER

FRAMEBUFFER_INCOMPLETE_LAYER_TARGETS

FRAMEBUFFER_INCOMPLETE_MISSING_ATTACHMENT

FRAMEBUFFER_INCOMPLETE_MULTISAMPLE

FRAMEBUFFER_INCOMPLETE_READ_BUFFER

FRAMEBUFFER_SRGB

FRAMEBUFFER_UNDEFINED

FRAMEBUFFER_UNSUPPORTED

FRONT

FRONT_AND_BACK

FRONT_FACE

FRONT_LEFT

FRONT_RIGHT

FUNC_ADD

FUNC_REVERSE_SUBTRACT

FUNC_SUBTRACT

GENERATE_MIPMAP

GENERATE_MIPMAP_HINT

GEOMETRY_INPUT_TYPE

GEOMETRY_OUTPUT_TYPE

GEOMETRY_SHADER

GEOMETRY_VERTICES_OUT

GEQUAL

GPU_DISJOINT_EXT

GREATER

GREEN

GREEN_BIAS

GREEN_BITS

GREEN_INTEGER

GREEN_SCALE

HALF_FLOAT

HALF_FLOAT_OES

HARDLIGHT_KHR

HIGH_FLOAT

HIGH_INT

HINT_BIT

HSL_COLOR_KHR

HSL_HUE_KHR

HSL_LUMINOSITY_KHR

HSL_SATURATION_KHR

IMAGE_2D

IMAGE_2D_ARRAY

IMAGE_3D

IMAGE_BINDING_ACCESS

IMAGE_BINDING_FORMAT

IMAGE_BINDING_LAYER

IMAGE_BINDING_LAYERED

IMAGE_BINDING_LEVEL

IMAGE_BINDING_NAME

IMAGE_CUBE

IMAGE_FORMAT_COMPATIBILITY_BY_CLASS

IMAGE_FORMAT_COMPATIBILITY_BY_SIZE

IMAGE_FORMAT_COMPATIBILITY_TYPE

IMPLEMENTATION_COLOR_READ_FORMAT

IMPLEMENTATION_COLOR_READ_TYPE

INCR

INCR_WRAP

INDEX

INDEX_ARRAY

INDEX_ARRAY_BUFFER_BINDING

INDEX_ARRAY_POINTER

INDEX_ARRAY_STRIDE

INDEX_ARRAY_TYPE

INDEX_BITS

INDEX_CLEAR_VALUE

INDEX_LOGIC_OP

INDEX_MODE

INDEX_OFFSET

INDEX_SHIFT

INDEX_WRITEMASK

INFO_LOG_LENGTH

INT

INTENSITY

INTENSITY4

INTENSITY8

INTENSITY12

INTENSITY16

INTERLEAVED_ATTRIBS

INTERPOLATE

INT_2_10_10_10_REV

INT_IMAGE_2D

INT_IMAGE_2D_ARRAY

INT_IMAGE_3D

INT_IMAGE_CUBE

INT_SAMPLER_1D

INT_SAMPLER_1D_ARRAY

INT_SAMPLER_2D

INT_SAMPLER_2D_ARRAY

INT_SAMPLER_2D_MULTISAMPLE

INT_SAMPLER_2D_MULTISAMPLE_ARRAY

INT_SAMPLER_2D_RECT

INT_SAMPLER_3D

INT_SAMPLER_BUFFER

INT_SAMPLER_CUBE

INT_VEC2

INT_VEC3

INT_VEC4

INVALID_ENUM

INVALID_FRAMEBUFFER_OPERATION

INVALID_INDEX

INVALID_OPERATION

INVALID_VALUE

INVERT

IS_ROW_MAJOR

KEEP

LAST_VERTEX_CONVENTION

LEFT

LEQUAL

LESS

LIGHT0

LIGHT1

LIGHT2

LIGHT3

LIGHT4

LIGHT5

LIGHT6

LIGHT7

LIGHTEN_KHR

LIGHTING

LIGHTING_BIT

LIGHT_MODEL_AMBIENT

LIGHT_MODEL_COLOR_CONTROL

LIGHT_MODEL_LOCAL_VIEWER

LIGHT_MODEL_TWO_SIDE

LINE

LINEAR

LINEAR_ATTENUATION

LINEAR_MIPMAP_LINEAR

LINEAR_MIPMAP_NEAREST

LINES

LINES_ADJACENCY

LINE_BIT

LINE_LOOP

LINE_RESET_TOKEN

LINE_SMOOTH

LINE_SMOOTH_HINT

LINE_STIPPLE

LINE_STIPPLE_PATTERN

LINE_STIPPLE_REPEAT

LINE_STRIP

LINE_STRIP_ADJACENCY

LINE_TOKEN

LINE_WIDTH

LINE_WIDTH_GRANULARITY

LINE_WIDTH_RANGE

LINK_STATUS

LIST_BASE

LIST_BIT

LIST_INDEX

LIST_MODE

LOAD

LOCATION

LOGIC_OP

LOGIC_OP_MODE

LOWER_LEFT

LOW_FLOAT

LOW_INT

LUMINANCE

LUMINANCE4

LUMINANCE4_ALPHA4

LUMINANCE6_ALPHA2

LUMINANCE8

LUMINANCE8_ALPHA8

LUMINANCE8_ALPHA8_EXT

LUMINANCE8_EXT

LUMINANCE12

LUMINANCE12_ALPHA4

LUMINANCE12_ALPHA12

LUMINANCE16

LUMINANCE16F_EXT

LUMINANCE16_ALPHA16

LUMINANCE32F_EXT

LUMINANCE_ALPHA

LUMINANCE_ALPHA16F_EXT

LUMINANCE_ALPHA32F_EXT

MAJOR_VERSION

MAP1_COLOR_4

MAP1_GRID_DOMAIN

MAP1_GRID_SEGMENTS

MAP1_INDEX

MAP1_NORMAL

MAP1_TEXTURE_COORD_1

MAP1_TEXTURE_COORD_2

MAP1_TEXTURE_COORD_3

MAP1_TEXTURE_COORD_4

MAP1_VERTEX_3

MAP1_VERTEX_4

MAP2_COLOR_4

MAP2_GRID_DOMAIN

MAP2_GRID_SEGMENTS

MAP2_INDEX

MAP2_NORMAL

MAP2_TEXTURE_COORD_1

MAP2_TEXTURE_COORD_2

MAP2_TEXTURE_COORD_3

MAP2_TEXTURE_COORD_4

MAP2_VERTEX_3

MAP2_VERTEX_4

MAP_COHERENT_BIT

MAP_COHERENT_BIT_EXT

MAP_COLOR

MAP_FLUSH_EXPLICIT_BIT

MAP_INVALIDATE_BUFFER_BIT

MAP_INVALIDATE_RANGE_BIT

MAP_PERSISTENT_BIT

MAP_PERSISTENT_BIT_EXT

MAP_READ_BIT

MAP_STENCIL

MAP_UNSYNCHRONIZED_BIT

MAP_WRITE_BIT

MATRIX_MODE

MATRIX_STRIDE

MAX

MAX_3D_TEXTURE_SIZE

MAX_ARRAY_TEXTURE_LAYERS

MAX_ATOMIC_COUNTER_BUFFER_BINDINGS

MAX_ATOMIC_COUNTER_BUFFER_SIZE

MAX_ATTRIB_STACK_DEPTH

MAX_CLIENT_ATTRIB_STACK_DEPTH

MAX_CLIP_DISTANCES

MAX_CLIP_PLANES

MAX_COLOR_ATTACHMENTS

MAX_COLOR_TEXTURE_SAMPLES

MAX_COMBINED_ATOMIC_COUNTERS

MAX_COMBINED_ATOMIC_COUNTER_BUFFERS

MAX_COMBINED_COMPUTE_UNIFORM_COMPONENTS

MAX_COMBINED_FRAGMENT_UNIFORM_COMPONENTS

MAX_COMBINED_GEOMETRY_UNIFORM_COMPONENTS

MAX_COMBINED_IMAGE_UNIFORMS

MAX_COMBINED_IMAGE_UNITS_AND_FRAGMENT_OUTPUTS

MAX_COMBINED_SHADER_OUTPUT_RESOURCES

MAX_COMBINED_SHADER_STORAGE_BLOCKS

MAX_COMBINED_TEXTURE_IMAGE_UNITS

MAX_COMBINED_UNIFORM_BLOCKS

MAX_COMBINED_VERTEX_UNIFORM_COMPONENTS

MAX_COMPUTE_ATOMIC_COUNTERS

MAX_COMPUTE_ATOMIC_COUNTER_BUFFERS

MAX_COMPUTE_IMAGE_UNIFORMS

MAX_COMPUTE_SHADER_STORAGE_BLOCKS

MAX_COMPUTE_SHARED_MEMORY_SIZE

MAX_COMPUTE_TEXTURE_IMAGE_UNITS

MAX_COMPUTE_UNIFORM_BLOCKS

MAX_COMPUTE_UNIFORM_COMPONENTS

MAX_COMPUTE_WORK_GROUP_COUNT

MAX_COMPUTE_WORK_GROUP_INVOCATIONS

MAX_COMPUTE_WORK_GROUP_SIZE

MAX_CUBE_MAP_TEXTURE_SIZE

MAX_DEBUG_GROUP_STACK_DEPTH

MAX_DEBUG_GROUP_STACK_DEPTH_KHR

MAX_DEBUG_LOGGED_MESSAGES

MAX_DEBUG_LOGGED_MESSAGES_KHR

MAX_DEBUG_MESSAGE_LENGTH

MAX_DEBUG_MESSAGE_LENGTH_KHR

MAX_DEPTH_TEXTURE_SAMPLES

MAX_DRAW_BUFFERS

MAX_DUAL_SOURCE_DRAW_BUFFERS

MAX_ELEMENTS_INDICES

MAX_ELEMENTS_VERTICES

MAX_ELEMENT_INDEX

MAX_EVAL_ORDER

MAX_FRAGMENT_ATOMIC_COUNTERS

MAX_FRAGMENT_ATOMIC_COUNTER_BUFFERS

MAX_FRAGMENT_IMAGE_UNIFORMS

MAX_FRAGMENT_INPUT_COMPONENTS

MAX_FRAGMENT_SHADER_STORAGE_BLOCKS

MAX_FRAGMENT_UNIFORM_BLOCKS

MAX_FRAGMENT_UNIFORM_COMPONENTS

MAX_FRAGMENT_UNIFORM_VECTORS

MAX_FRAMEBUFFER_HEIGHT

MAX_FRAMEBUFFER_SAMPLES

MAX_FRAMEBUFFER_WIDTH

MAX_GEOMETRY_INPUT_COMPONENTS

MAX_GEOMETRY_OUTPUT_COMPONENTS

MAX_GEOMETRY_OUTPUT_VERTICES

MAX_GEOMETRY_SHADER_STORAGE_BLOCKS

MAX_GEOMETRY_TEXTURE_IMAGE_UNITS

MAX_GEOMETRY_TOTAL_OUTPUT_COMPONENTS

MAX_GEOMETRY_UNIFORM_BLOCKS

MAX_GEOMETRY_UNIFORM_COMPONENTS

MAX_IMAGE_UNITS

MAX_INTEGER_SAMPLES

MAX_LABEL_LENGTH

MAX_LABEL_LENGTH_KHR

MAX_LIGHTS

MAX_LIST_NESTING

MAX_MODELVIEW_STACK_DEPTH

MAX_NAME_LENGTH

MAX_NAME_STACK_DEPTH

MAX_NUM_ACTIVE_VARIABLES

MAX_PIXEL_MAP_TABLE

MAX_PROGRAM_TEXEL_OFFSET

MAX_PROGRAM_TEXTURE_GATHER_OFFSET

MAX_PROJECTION_STACK_DEPTH

MAX_RECTANGLE_TEXTURE_SIZE

MAX_RECTANGLE_TEXTURE_SIZE_ARB

MAX_RENDERBUFFER_SIZE

MAX_SAMPLES

MAX_SAMPLE_MASK_WORDS

MAX_SERVER_WAIT_TIMEOUT

MAX_SHADER_PIXEL_LOCAL_STORAGE_FAST_SIZE_EXT

MAX_SHADER_PIXEL_LOCAL_STORAGE_SIZE_EXT

MAX_SHADER_STORAGE_BLOCK_SIZE

MAX_SHADER_STORAGE_BUFFER_BINDINGS

MAX_TESS_CONTROL_SHADER_STORAGE_BLOCKS

MAX_TESS_EVALUATION_SHADER_STORAGE_BLOCKS

MAX_TEXTURE_BUFFER_SIZE

MAX_TEXTURE_COORDS

MAX_TEXTURE_IMAGE_UNITS

MAX_TEXTURE_LOD_BIAS

MAX_TEXTURE_MAX_ANISOTROPY_EXT

MAX_TEXTURE_SIZE

MAX_TEXTURE_STACK_DEPTH

MAX_TEXTURE_UNITS

MAX_TRANSFORM_FEEDBACK_INTERLEAVED_COMPONENTS

MAX_TRANSFORM_FEEDBACK_SEPARATE_ATTRIBS

MAX_TRANSFORM_FEEDBACK_SEPARATE_COMPONENTS

MAX_UNIFORM_BLOCK_SIZE

MAX_UNIFORM_BUFFER_BINDINGS

MAX_UNIFORM_LOCATIONS

MAX_VARYING_COMPONENTS

MAX_VARYING_FLOATS

MAX_VARYING_VECTORS

MAX_VERTEX_ATOMIC_COUNTERS

MAX_VERTEX_ATOMIC_COUNTER_BUFFERS

MAX_VERTEX_ATTRIBS

MAX_VERTEX_ATTRIB_BINDINGS

MAX_VERTEX_ATTRIB_RELATIVE_OFFSET

MAX_VERTEX_ATTRIB_STRIDE

MAX_VERTEX_IMAGE_UNIFORMS

MAX_VERTEX_OUTPUT_COMPONENTS

MAX_VERTEX_SHADER_STORAGE_BLOCKS

MAX_VERTEX_TEXTURE_IMAGE_UNITS

MAX_VERTEX_UNIFORM_BLOCKS

MAX_VERTEX_UNIFORM_COMPONENTS

MAX_VERTEX_UNIFORM_VECTORS

MAX_VIEWPORT_DIMS

MEDIUM_FLOAT

MEDIUM_INT

MIN

MINOR_VERSION

MIN_PROGRAM_TEXEL_OFFSET

MIN_PROGRAM_TEXTURE_GATHER_OFFSET

MIRRORED_REPEAT

MODELVIEW

MODELVIEW_MATRIX

MODELVIEW_STACK_DEPTH

MODULATE

MULT

MULTIPLY_KHR

MULTISAMPLE

MULTISAMPLE_BIT

MULTISAMPLE_BUFFER_BIT0_QCOM

MULTISAMPLE_BUFFER_BIT1_QCOM

MULTISAMPLE_BUFFER_BIT2_QCOM

MULTISAMPLE_BUFFER_BIT3_QCOM

MULTISAMPLE_BUFFER_BIT4_QCOM

MULTISAMPLE_BUFFER_BIT5_QCOM

MULTISAMPLE_BUFFER_BIT6_QCOM

MULTISAMPLE_BUFFER_BIT7_QCOM

N3F_V3F

NAME_LENGTH

NAME_STACK_DEPTH

NAND

NEAREST

NEAREST_MIPMAP_LINEAR

NEAREST_MIPMAP_NEAREST

NEVER

NICEST

NONE

NOOP

NOR

NORMALIZE

NORMAL_ARRAY

NORMAL_ARRAY_BUFFER_BINDING

NORMAL_ARRAY_POINTER

NORMAL_ARRAY_STRIDE

NORMAL_ARRAY_TYPE

NORMAL_MAP

NOTEQUAL

NO_ERROR

NUM_ACTIVE_VARIABLES

NUM_COMPRESSED_TEXTURE_FORMATS

NUM_EXTENSIONS

NUM_PROGRAM_BINARY_FORMATS

NUM_SAMPLE_COUNTS

NUM_SHADER_BINARY_FORMATS

OBJECT_LINEAR

OBJECT_PLANE

OBJECT_TYPE

OFFSET

ONE

ONE_MINUS_CONSTANT_ALPHA

ONE_MINUS_CONSTANT_COLOR

ONE_MINUS_DST_ALPHA

ONE_MINUS_DST_COLOR

ONE_MINUS_SRC1_ALPHA

ONE_MINUS_SRC1_COLOR

ONE_MINUS_SRC_ALPHA

ONE_MINUS_SRC_COLOR

OPERAND0_ALPHA

OPERAND0_RGB

OPERAND1_ALPHA

OPERAND1_RGB

OPERAND2_ALPHA

OPERAND2_RGB

OR

ORDER

OR_INVERTED

OR_REVERSE

OUT_OF_MEMORY

OVERLAY_KHR

PACK_ALIGNMENT

PACK_IMAGE_HEIGHT

PACK_LSB_FIRST

PACK_ROW_LENGTH

PACK_SKIP_IMAGES

PACK_SKIP_PIXELS

PACK_SKIP_ROWS

PACK_SWAP_BYTES

PASS_THROUGH_TOKEN

PERSPECTIVE_CORRECTION_HINT

PIXEL_BUFFER_BARRIER_BIT

PIXEL_MAP_A_TO_A

PIXEL_MAP_A_TO_A_SIZE

PIXEL_MAP_B_TO_B

PIXEL_MAP_B_TO_B_SIZE

PIXEL_MAP_G_TO_G

PIXEL_MAP_G_TO_G_SIZE

PIXEL_MAP_I_TO_A

PIXEL_MAP_I_TO_A_SIZE

PIXEL_MAP_I_TO_B

PIXEL_MAP_I_TO_B_SIZE

PIXEL_MAP_I_TO_G

PIXEL_MAP_I_TO_G_SIZE

PIXEL_MAP_I_TO_I

PIXEL_MAP_I_TO_I_SIZE

PIXEL_MAP_I_TO_R

PIXEL_MAP_I_TO_R_SIZE

PIXEL_MAP_R_TO_R

PIXEL_MAP_R_TO_R_SIZE

PIXEL_MAP_S_TO_S

PIXEL_MAP_S_TO_S_SIZE

PIXEL_MODE_BIT

PIXEL_PACK_BUFFER

PIXEL_PACK_BUFFER_BINDING

PIXEL_UNPACK_BUFFER

PIXEL_UNPACK_BUFFER_BINDING

POINT

POINTS

POINT_BIT

POINT_DISTANCE_ATTENUATION

POINT_FADE_THRESHOLD_SIZE

POINT_SIZE

POINT_SIZE_GRANULARITY

POINT_SIZE_MAX

POINT_SIZE_MIN

POINT_SIZE_RANGE

POINT_SMOOTH

POINT_SMOOTH_HINT

POINT_SPRITE

POINT_SPRITE_COORD_ORIGIN

POINT_TOKEN

POLYGON

POLYGON_BIT

POLYGON_MODE

POLYGON_OFFSET_FACTOR

POLYGON_OFFSET_FILL

POLYGON_OFFSET_LINE

POLYGON_OFFSET_POINT

POLYGON_OFFSET_UNITS

POLYGON_SMOOTH

POLYGON_SMOOTH_HINT

POLYGON_STIPPLE

POLYGON_STIPPLE_BIT

POLYGON_TOKEN

POSITION

PREVIOUS

PRIMARY_COLOR

PRIMITIVES_GENERATED

PRIMITIVE_RESTART

PRIMITIVE_RESTART_FIXED_INDEX

PRIMITIVE_RESTART_INDEX

PROGRAM

PROGRAM_BINARY_FORMATS

PROGRAM_BINARY_LENGTH

PROGRAM_BINARY_RETRIEVABLE_HINT

PROGRAM_INPUT

PROGRAM_KHR

PROGRAM_OUTPUT

PROGRAM_PIPELINE

PROGRAM_PIPELINE_BINDING

PROGRAM_PIPELINE_KHR

PROGRAM_POINT_SIZE

PROGRAM_SEPARABLE

PROJECTION

PROJECTION_MATRIX

PROJECTION_STACK_DEPTH

PROVOKING_VERTEX

PROXY_TEXTURE_1D

PROXY_TEXTURE_1D_ARRAY

PROXY_TEXTURE_2D

PROXY_TEXTURE_2D_ARRAY

PROXY_TEXTURE_2D_MULTISAMPLE

PROXY_TEXTURE_2D_MULTISAMPLE_ARRAY

PROXY_TEXTURE_3D

PROXY_TEXTURE_CUBE_MAP

PROXY_TEXTURE_RECTANGLE

PROXY_TEXTURE_RECTANGLE_ARB

Q

QUADRATIC_ATTENUATION

QUADS

QUADS_FOLLOW_PROVOKING_VERTEX_CONVENTION

QUAD_STRIP

QUERY

QUERY_BY_REGION_NO_WAIT

QUERY_BY_REGION_WAIT

QUERY_COUNTER_BITS

QUERY_COUNTER_BITS_EXT

QUERY_KHR

QUERY_NO_WAIT

QUERY_RESULT

QUERY_RESULT_AVAILABLE

QUERY_RESULT_AVAILABLE_EXT

QUERY_RESULT_EXT

QUERY_WAIT

R

R3_G3_B2

R8

R8I

R8UI

R8_EXT

R8_SNORM

R11F_G11F_B10F

R16

R16F

R16F_EXT

R16I

R16UI

R16_SNORM

R32F

R32F_EXT

R32I

R32UI

RASTERIZER_DISCARD

READ_BUFFER

READ_FRAMEBUFFER

READ_FRAMEBUFFER_BINDING

READ_ONLY

READ_WRITE

RED

RED_BIAS

RED_BITS

RED_INTEGER

RED_SCALE

REFERENCED_BY_COMPUTE_SHADER

REFERENCED_BY_FRAGMENT_SHADER

REFERENCED_BY_VERTEX_SHADER

REFLECTION_MAP

RENDER

RENDERBUFFER

RENDERBUFFER_ALPHA_SIZE

RENDERBUFFER_BINDING

RENDERBUFFER_BLUE_SIZE

RENDERBUFFER_DEPTH_SIZE

RENDERBUFFER_GREEN_SIZE

RENDERBUFFER_HEIGHT

RENDERBUFFER_INTERNAL_FORMAT

RENDERBUFFER_RED_SIZE

RENDERBUFFER_SAMPLES

RENDERBUFFER_STENCIL_SIZE

RENDERBUFFER_WIDTH

RENDERER

RENDER_MODE

REPEAT

REPLACE

REQUIRED_TEXTURE_IMAGE_UNITS_OES

RESCALE_NORMAL

RETURN

RG

RG8

RG8I

RG8UI

RG8_EXT

RG8_SNORM

RG16

RG16F

RG16F_EXT

RG16I

RG16UI

RG16_SNORM

RG32F

RG32F_EXT

RG32I

RG32UI

RGB

RGB4

RGB5

RGB5_A1

RGB8

RGB8I

RGB8UI

RGB8_SNORM

RGB9_E5

RGB10

RGB10_A2

RGB10_A2UI

RGB10_A2_EXT

RGB10_EXT

RGB12

RGB16

RGB16F

RGB16F_EXT

RGB16I

RGB16UI

RGB16_SNORM

RGB32F

RGB32F_EXT

RGB32I

RGB32UI

RGB565

RGBA

RGBA2

RGBA4

RGBA8

RGBA8I

RGBA8UI

RGBA8_SNORM

RGBA12

RGBA16

RGBA16F

RGBA16F_EXT

RGBA16I

RGBA16UI

RGBA16_SNORM

RGBA32F

RGBA32F_EXT

RGBA32I

RGBA32UI

RGBA_INTEGER

RGBA_MODE

RGB_INTEGER

RGB_SCALE

RG_INTEGER

RIGHT

S

SAMPLER

SAMPLER_1D

SAMPLER_1D_ARRAY

SAMPLER_1D_ARRAY_SHADOW

SAMPLER_1D_SHADOW

SAMPLER_2D

SAMPLER_2D_ARRAY

SAMPLER_2D_ARRAY_SHADOW

SAMPLER_2D_MULTISAMPLE

SAMPLER_2D_MULTISAMPLE_ARRAY

SAMPLER_2D_RECT

SAMPLER_2D_RECT_SHADOW

SAMPLER_2D_SHADOW

SAMPLER_3D

SAMPLER_BINDING

SAMPLER_BUFFER

SAMPLER_CUBE

SAMPLER_CUBE_SHADOW

SAMPLER_EXTERNAL_OES

SAMPLER_KHR

SAMPLES

SAMPLES_PASSED

SAMPLE_ALPHA_TO_COVERAGE

SAMPLE_ALPHA_TO_ONE

SAMPLE_BUFFERS

SAMPLE_COVERAGE

SAMPLE_COVERAGE_INVERT

SAMPLE_COVERAGE_VALUE

SAMPLE_MASK

SAMPLE_MASK_VALUE

SAMPLE_POSITION

SCISSOR_BIT

SCISSOR_BOX

SCISSOR_TEST

SCREEN_KHR

SECONDARY_COLOR_ARRAY

SECONDARY_COLOR_ARRAY_BUFFER_BINDING

SECONDARY_COLOR_ARRAY_POINTER

SECONDARY_COLOR_ARRAY_SIZE

SECONDARY_COLOR_ARRAY_STRIDE

SECONDARY_COLOR_ARRAY_TYPE

SELECT

SELECTION_BUFFER_POINTER

SELECTION_BUFFER_SIZE

SEPARATE_ATTRIBS

SEPARATE_SPECULAR_COLOR

SET

SHADER

SHADER_BINARY_FORMATS

SHADER_COMPILER

SHADER_IMAGE_ACCESS_BARRIER_BIT

SHADER_KHR

SHADER_PIXEL_LOCAL_STORAGE_EXT

SHADER_SOURCE_LENGTH

SHADER_STORAGE_BARRIER_BIT

SHADER_STORAGE_BLOCK

SHADER_STORAGE_BUFFER

SHADER_STORAGE_BUFFER_BINDING

SHADER_STORAGE_BUFFER_OFFSET_ALIGNMENT

SHADER_STORAGE_BUFFER_SIZE

SHADER_STORAGE_BUFFER_START

SHADER_TYPE

SHADE_MODEL

SHADING_LANGUAGE_VERSION

SHININESS

SHORT

SIGNALED

SIGNED_NORMALIZED

SINGLE_COLOR

SLUMINANCE

SLUMINANCE8

SLUMINANCE8_ALPHA8

SLUMINANCE_ALPHA

SMOOTH

SMOOTH_LINE_WIDTH_GRANULARITY

SMOOTH_LINE_WIDTH_RANGE

SMOOTH_POINT_SIZE_GRANULARITY

SMOOTH_POINT_SIZE_RANGE

SOFTLIGHT_KHR

SOURCE0_ALPHA

SOURCE0_RGB

SOURCE1_ALPHA

SOURCE1_RGB

SOURCE2_ALPHA

SOURCE2_RGB

SPECULAR

SPHERE_MAP

SPOT_CUTOFF

SPOT_DIRECTION

SPOT_EXPONENT

SRC0_ALPHA

SRC0_RGB

SRC1_ALPHA

SRC1_COLOR

SRC1_RGB

SRC2_ALPHA

SRC2_RGB

SRC_ALPHA

SRC_ALPHA_SATURATE

SRC_COLOR

SRGB

SRGB8

SRGB8_ALPHA8

SRGB_ALPHA

STACK_OVERFLOW

STACK_OVERFLOW_KHR

STACK_UNDERFLOW

STACK_UNDERFLOW_KHR

STATIC_COPY

STATIC_DRAW

STATIC_READ

STENCIL

STENCIL_ATTACHMENT

STENCIL_BACK_FAIL

STENCIL_BACK_FUNC

STENCIL_BACK_PASS_DEPTH_FAIL

STENCIL_BACK_PASS_DEPTH_PASS

STENCIL_BACK_REF

STENCIL_BACK_VALUE_MASK

STENCIL_BACK_WRITEMASK

STENCIL_BITS

STENCIL_BUFFER_BIT

STENCIL_BUFFER_BIT0_QCOM

STENCIL_BUFFER_BIT1_QCOM

STENCIL_BUFFER_BIT2_QCOM

STENCIL_BUFFER_BIT3_QCOM

STENCIL_BUFFER_BIT4_QCOM

STENCIL_BUFFER_BIT5_QCOM

STENCIL_BUFFER_BIT6_QCOM

STENCIL_BUFFER_BIT7_QCOM

STENCIL_CLEAR_VALUE

STENCIL_FAIL

STENCIL_FUNC

STENCIL_INDEX

STENCIL_INDEX1

STENCIL_INDEX4

STENCIL_INDEX8

STENCIL_INDEX16

STENCIL_PASS_DEPTH_FAIL

STENCIL_PASS_DEPTH_PASS

STENCIL_REF

STENCIL_TEST

STENCIL_VALUE_MASK

STENCIL_WRITEMASK

STEREO

STORAGE_CACHED_APPLE

STORAGE_PRIVATE_APPLE

STORAGE_SHARED_APPLE

STREAM_COPY

STREAM_DRAW

STREAM_READ

SUBPIXEL_BITS

SUBTRACT

SYNC_CONDITION

SYNC_FENCE

SYNC_FLAGS

SYNC_FLUSH_COMMANDS_BIT

SYNC_GPU_COMMANDS_COMPLETE

SYNC_STATUS

T

T2F_C3F_V3F

T2F_C4F_N3F_V3F

T2F_C4UB_V3F

T2F_N3F_V3F

T2F_V3F

T4F_C4F_N3F_V4F

T4F_V4F

TEXTURE

TEXTURE0

TEXTURE1

TEXTURE2

TEXTURE3

TEXTURE4

TEXTURE5

TEXTURE6

TEXTURE7

TEXTURE8

TEXTURE9

TEXTURE10

TEXTURE11

TEXTURE12

TEXTURE13

TEXTURE14

TEXTURE15

TEXTURE16

TEXTURE17

TEXTURE18

TEXTURE19

TEXTURE20

TEXTURE21

TEXTURE22

TEXTURE23

TEXTURE24

TEXTURE25

TEXTURE26

TEXTURE27

TEXTURE28

TEXTURE29

TEXTURE30

TEXTURE31

TEXTURE_1D

TEXTURE_1D_ARRAY

TEXTURE_2D

TEXTURE_2D_ARRAY

TEXTURE_2D_MULTISAMPLE

TEXTURE_2D_MULTISAMPLE_ARRAY

TEXTURE_3D

TEXTURE_ALPHA_SIZE

TEXTURE_ALPHA_TYPE

TEXTURE_BASE_LEVEL

TEXTURE_BINDING_1D

TEXTURE_BINDING_1D_ARRAY

TEXTURE_BINDING_2D

TEXTURE_BINDING_2D_ARRAY

TEXTURE_BINDING_2D_MULTISAMPLE

TEXTURE_BINDING_2D_MULTISAMPLE_ARRAY

TEXTURE_BINDING_3D

TEXTURE_BINDING_BUFFER

TEXTURE_BINDING_CUBE_MAP

TEXTURE_BINDING_EXTERNAL_OES

TEXTURE_BINDING_RECTANGLE

TEXTURE_BINDING_RECTANGLE_ARB

TEXTURE_BIT

TEXTURE_BLUE_SIZE

TEXTURE_BLUE_TYPE

TEXTURE_BORDER

TEXTURE_BORDER_COLOR

TEXTURE_BUFFER

TEXTURE_BUFFER_DATA_STORE_BINDING

TEXTURE_COMPARE_FUNC

TEXTURE_COMPARE_MODE

TEXTURE_COMPONENTS

TEXTURE_COMPRESSED

TEXTURE_COMPRESSED_IMAGE_SIZE

TEXTURE_COMPRESSION_HINT

TEXTURE_COORD_ARRAY

TEXTURE_COORD_ARRAY_BUFFER_BINDING

TEXTURE_COORD_ARRAY_POINTER

TEXTURE_COORD_ARRAY_SIZE

TEXTURE_COORD_ARRAY_STRIDE

TEXTURE_COORD_ARRAY_TYPE

TEXTURE_CUBE_MAP

TEXTURE_CUBE_MAP_NEGATIVE_X

TEXTURE_CUBE_MAP_NEGATIVE_Y

TEXTURE_CUBE_MAP_NEGATIVE_Z

TEXTURE_CUBE_MAP_POSITIVE_X

TEXTURE_CUBE_MAP_POSITIVE_Y

TEXTURE_CUBE_MAP_POSITIVE_Z

TEXTURE_CUBE_MAP_SEAMLESS

TEXTURE_DEPTH

TEXTURE_DEPTH_SIZE

TEXTURE_DEPTH_TYPE

TEXTURE_ENV

TEXTURE_ENV_COLOR

TEXTURE_ENV_MODE

TEXTURE_EXTERNAL_OES

TEXTURE_FETCH_BARRIER_BIT

TEXTURE_FILTER_CONTROL

TEXTURE_FIXED_SAMPLE_LOCATIONS

TEXTURE_GEN_MODE

TEXTURE_GEN_Q

TEXTURE_GEN_R

TEXTURE_GEN_S

TEXTURE_GEN_T

TEXTURE_GREEN_SIZE

TEXTURE_GREEN_TYPE

TEXTURE_HEIGHT

TEXTURE_IMMUTABLE_FORMAT

TEXTURE_IMMUTABLE_FORMAT_EXT

TEXTURE_IMMUTABLE_LEVELS

TEXTURE_INTENSITY_SIZE

TEXTURE_INTENSITY_TYPE

TEXTURE_INTERNAL_FORMAT

TEXTURE_LOD_BIAS

TEXTURE_LUMINANCE_SIZE

TEXTURE_LUMINANCE_TYPE

TEXTURE_MAG_FILTER

TEXTURE_MATRIX

TEXTURE_MAX_ANISOTROPY_EXT

TEXTURE_MAX_LEVEL

TEXTURE_MAX_LOD

TEXTURE_MIN_FILTER

TEXTURE_MIN_LOD

TEXTURE_PRIORITY

TEXTURE_RANGE_LENGTH_APPLE

TEXTURE_RANGE_POINTER_APPLE

TEXTURE_RECTANGLE

TEXTURE_RECTANGLE_ARB

TEXTURE_RED_SIZE

TEXTURE_RED_TYPE

TEXTURE_RESIDENT

TEXTURE_SAMPLES

TEXTURE_SHARED_SIZE

TEXTURE_STACK_DEPTH

TEXTURE_STENCIL_SIZE

TEXTURE_STORAGE_HINT_APPLE

TEXTURE_SWIZZLE_A

TEXTURE_SWIZZLE_B

TEXTURE_SWIZZLE_G

TEXTURE_SWIZZLE_R

TEXTURE_SWIZZLE_RGBA

TEXTURE_UPDATE_BARRIER_BIT

TEXTURE_USAGE_ANGLE

TEXTURE_WIDTH

TEXTURE_WRAP_R

TEXTURE_WRAP_S

TEXTURE_WRAP_T

TIMEOUT_EXPIRED

TIMEOUT_IGNORED

TIMESTAMP

TIMESTAMP_EXT

TIME_ELAPSED

TIME_ELAPSED_EXT

TOP_LEVEL_ARRAY_SIZE

TOP_LEVEL_ARRAY_STRIDE

TRANSFORM_BIT

TRANSFORM_FEEDBACK

TRANSFORM_FEEDBACK_ACTIVE

TRANSFORM_FEEDBACK_BARRIER_BIT

TRANSFORM_FEEDBACK_BINDING

TRANSFORM_FEEDBACK_BUFFER

TRANSFORM_FEEDBACK_BUFFER_BINDING

TRANSFORM_FEEDBACK_BUFFER_MODE

TRANSFORM_FEEDBACK_BUFFER_SIZE

TRANSFORM_FEEDBACK_BUFFER_START

TRANSFORM_FEEDBACK_PAUSED

TRANSFORM_FEEDBACK_PRIMITIVES_WRITTEN

TRANSFORM_FEEDBACK_VARYING

TRANSFORM_FEEDBACK_VARYINGS

TRANSFORM_FEEDBACK_VARYING_MAX_LENGTH

TRANSPOSE_COLOR_MATRIX

TRANSPOSE_MODELVIEW_MATRIX

TRANSPOSE_PROJECTION_MATRIX

TRANSPOSE_TEXTURE_MATRIX

TRIANGLES

TRIANGLES_ADJACENCY

TRIANGLE_FAN

TRIANGLE_STRIP

TRIANGLE_STRIP_ADJACENCY

TRUE

TYPE

UNIFORM

UNIFORM_ARRAY_STRIDE

UNIFORM_BARRIER_BIT

UNIFORM_BLOCK

UNIFORM_BLOCK_ACTIVE_UNIFORMS

UNIFORM_BLOCK_ACTIVE_UNIFORM_INDICES

UNIFORM_BLOCK_BINDING

UNIFORM_BLOCK_DATA_SIZE

UNIFORM_BLOCK_INDEX

UNIFORM_BLOCK_NAME_LENGTH

UNIFORM_BLOCK_REFERENCED_BY_FRAGMENT_SHADER

UNIFORM_BLOCK_REFERENCED_BY_GEOMETRY_SHADER

UNIFORM_BLOCK_REFERENCED_BY_VERTEX_SHADER

UNIFORM_BUFFER

UNIFORM_BUFFER_BINDING

UNIFORM_BUFFER_OFFSET_ALIGNMENT

UNIFORM_BUFFER_SIZE

UNIFORM_BUFFER_START

UNIFORM_IS_ROW_MAJOR

UNIFORM_MATRIX_STRIDE

UNIFORM_NAME_LENGTH

UNIFORM_OFFSET

UNIFORM_SIZE

UNIFORM_TYPE

UNPACK_ALIGNMENT

UNPACK_CLIENT_STORAGE_APPLE

UNPACK_IMAGE_HEIGHT

UNPACK_LSB_FIRST

UNPACK_ROW_LENGTH

UNPACK_SKIP_IMAGES

UNPACK_SKIP_PIXELS

UNPACK_SKIP_ROWS

UNPACK_SWAP_BYTES

UNSIGNALED

UNSIGNED_BYTE

UNSIGNED_BYTE_2_3_3_REV

UNSIGNED_BYTE_3_3_2

UNSIGNED_INT

UNSIGNED_INT_2_10_10_10_REV

UNSIGNED_INT_5_9_9_9_REV

UNSIGNED_INT_8_8_8_8

UNSIGNED_INT_8_8_8_8_REV

UNSIGNED_INT_10F_11F_11F_REV

UNSIGNED_INT_10_10_10_2

UNSIGNED_INT_24_8

UNSIGNED_INT_ATOMIC_COUNTER

UNSIGNED_INT_IMAGE_2D

UNSIGNED_INT_IMAGE_2D_ARRAY

UNSIGNED_INT_IMAGE_3D

UNSIGNED_INT_IMAGE_CUBE

UNSIGNED_INT_SAMPLER_1D

UNSIGNED_INT_SAMPLER_1D_ARRAY

UNSIGNED_INT_SAMPLER_2D

UNSIGNED_INT_SAMPLER_2D_ARRAY

UNSIGNED_INT_SAMPLER_2D_MULTISAMPLE

UNSIGNED_INT_SAMPLER_2D_MULTISAMPLE_ARRAY

UNSIGNED_INT_SAMPLER_2D_RECT

UNSIGNED_INT_SAMPLER_3D

UNSIGNED_INT_SAMPLER_BUFFER

UNSIGNED_INT_SAMPLER_CUBE

UNSIGNED_INT_VEC2

UNSIGNED_INT_VEC3

UNSIGNED_INT_VEC4

UNSIGNED_NORMALIZED

UNSIGNED_SHORT

UNSIGNED_SHORT_1_5_5_5_REV

UNSIGNED_SHORT_4_4_4_4

UNSIGNED_SHORT_4_4_4_4_REV

UNSIGNED_SHORT_5_5_5_1

UNSIGNED_SHORT_5_6_5

UNSIGNED_SHORT_5_6_5_REV

UPPER_LEFT

V2F

V3F

VALIDATE_STATUS

VENDOR

VERSION

VERTEX_ARRAY

VERTEX_ARRAY_BINDING

VERTEX_ARRAY_BINDING_APPLE

VERTEX_ARRAY_BUFFER_BINDING

VERTEX_ARRAY_KHR

VERTEX_ARRAY_POINTER

VERTEX_ARRAY_SIZE

VERTEX_ARRAY_STRIDE

VERTEX_ARRAY_TYPE

VERTEX_ATTRIB_ARRAY_BARRIER_BIT

VERTEX_ATTRIB_ARRAY_BUFFER_BINDING

VERTEX_ATTRIB_ARRAY_DIVISOR

VERTEX_ATTRIB_ARRAY_ENABLED

VERTEX_ATTRIB_ARRAY_INTEGER

VERTEX_ATTRIB_ARRAY_NORMALIZED

VERTEX_ATTRIB_ARRAY_POINTER

VERTEX_ATTRIB_ARRAY_SIZE

VERTEX_ATTRIB_ARRAY_STRIDE

VERTEX_ATTRIB_ARRAY_TYPE

VERTEX_ATTRIB_BINDING

VERTEX_ATTRIB_RELATIVE_OFFSET

VERTEX_BINDING_BUFFER

VERTEX_BINDING_DIVISOR

VERTEX_BINDING_OFFSET

VERTEX_BINDING_STRIDE

VERTEX_PROGRAM_POINT_SIZE

VERTEX_PROGRAM_TWO_SIDE

VERTEX_SHADER

VERTEX_SHADER_BIT

VIEWPORT

VIEWPORT_BIT

WAIT_FAILED

WEIGHT_ARRAY_BUFFER_BINDING

WRITE_ONLY

XOR

ZERO

ZOOM_X

ZOOM_Y

Traits

Gl

Functions

bpp 🔒

buffer_data

buffer_data_raw

buffer_sub_data

calculate_length 🔒

Type Aliases

GLDEBUGPROC

GLDEBUGPROCAMD

GLDEBUGPROCARB

GLDEBUGPROCKHR

GLbitfield

GLboolean

GLbyte

GLchar

GLcharARB

GLclampd

GLclampf

GLclampx

GLdouble

GLeglImageOES

GLenum

GLfixed

GLfloat

GLhalf

GLhalfARB

GLhalfNV

GLhandleARB

GLint

GLint64

GLint64EXT

GLintptr

GLintptrARB

GLshort

GLsizei

GLsizeiptr

GLsizeiptrARB

GLsync

GLubyte

GLuint

GLuint64

GLuint64EXT

GLushort

GLvdpauSurfaceNV

GLvoid